+++ cypress/locators/badge/locators.js
@@ -1,0 +1,3 @@
+// component preview locators
+export const BADGE = '[data-component="badge"]';
+export const BADGE_COUNTER = '[data-element="badge-counter"]';