+++ cypress/support/step-definitions/pages-steps.js
@@ -13,7 +13,7 @@ When('I go to {word} page', (word) => {
 });
 
 When('I close page', () => {
-  closeDataElement().click();
+  closeDataElement().click({ multiple: true });
   cy.wait(500, { log: DEBUG_FLAG }); // wait was added due to changing animation
 });