+++ src/components/button/button.component.js
@@ -5,7 +5,6 @@ import Icon from '../icon';
 import StyledButton, { StyledButtonSubtext } from './button.style';
 import tagComponent from '../../utils/helpers/tags';
 import OptionsHelper from '../../utils/helpers/options-helper';
-import Logger from '../../utils/logger';
 
 const Button = (props) => {
   const {
@@ -14,18 +13,9 @@ const Button = (props) => {
 
   const { as, buttonType, ...rest } = props;
 
-  const IS_USING_DEPRECATED_TYPE_DESTRUCTIVE = buttonType === 'destructive';
-
-  if (IS_USING_DEPRECATED_TYPE_DESTRUCTIVE) {
-    Logger.deprecate(
-      'buttonType="destructive" has been deprecated. See https://github.com/Sage/carbon/releases for details.'
-    );
-  }
-
   const propsWithoutAs = {
     ...rest,
-    buttonType: buttonType || as,
-    ...(IS_USING_DEPRECATED_TYPE_DESTRUCTIVE && { buttonType: 'primary', destructive: true })
+    buttonType: buttonType || as
   };
 
   if (subtext.length > 0 && size !== 'large') {
@@ -52,6 +42,7 @@ function renderStyledButton(buttonProps) {
     theme,
     forwardRef,
     href,
+    styleOverride,
     ...styleProps
   } = buttonProps;
 
@@ -67,11 +58,11 @@ function renderStyledButton(buttonProps) {
       disabled={ disabled }
       role='button'
       type='button'
-      legacyColorVariant={ theme }
       iconType={ iconType }
       { ...tagComponent('button', buttonProps) }
       { ...styleProps }
       ref={ forwardRef }
+      styleOverride={ styleOverride }
     >
       { renderChildren(buttonProps) }
     </StyledButton>
@@ -85,7 +76,6 @@ function renderChildren({
     primary: 'on-dark-background',
     secondary: 'business-color',
     tertiary: 'business-color',
-    destructive: 'on-dark-background',
     darkBackground: 'business-color'
   };
 
@@ -100,7 +90,13 @@ function renderChildren({
         />) }
       <span>
         <span data-element='main-text'>{ children }</span>
-        { size === 'large' && <StyledButtonSubtext data-element='subtext'>{ subtext }</StyledButtonSubtext> }
+        { size === 'large' && (
+          <StyledButtonSubtext
+            data-element='subtext'
+          >
+            { subtext }
+          </StyledButtonSubtext>
+        )}
       </span>
       { iconType && iconPosition === 'after' && (
         <Icon
@@ -134,14 +130,15 @@ Button.propTypes = {
   forwardRef: PropTypes.func,
   /** Button types for legacy theme: "primary" | "secondary" */
   as: PropTypes.oneOf(OptionsHelper.themesBinary),
-  /** Set this prop to pass in legacy theme color variants */
-  theme: PropTypes.oneOf(OptionsHelper.buttonColors),
-  checkTheme: PropTypes.func,
-  /** Legacy - used to transfrom button into anchor */
+  /** Legacy - used to transform button into anchor */
   href: PropTypes.string,
-  /** Legacy - used to transfrom button into anchor */
-  to: PropTypes.string
-
+  /** Legacy - used to transform button into anchor */
+  to: PropTypes.string,
+  /** Allows override of existing component styles */
+  styleOverride: PropTypes.shape({
+    root: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+    icon: PropTypes.oneOfType([PropTypes.func, PropTypes.object])
+  })
 };
 
 Button.defaultProps = {
@@ -150,8 +147,8 @@ Button.defaultProps = {
   disabled: false,
   destructive: false,
   iconPosition: 'before',
-  theme: 'blue',
-  subtext: ''
+  subtext: '',
+  styleOverride: { root: {}, icon: {} }
 };
 
 const ButtonWithForwardRef = React.forwardRef((props, ref) => <Button forwardRef={ ref } { ...props } />);