+++ src/components/split-button/split-button-children.style.js
@@ -1,7 +1,6 @@
 import styled, { css } from 'styled-components';
 import BaseTheme from '../../style/themes/base';
 import StyledButton from '../button/button.style';
-import { getSplitButtonChildrenClassicStyles } from './split-button-classic.style';
 
 const StyledSplitButtonChildrenContainer = styled.div`
   ${({ theme, align }) => css`
@@ -34,8 +33,6 @@ const StyledSplitButtonChildrenContainer = styled.div`
       }
     }
   `}
-
-  ${getSplitButtonChildrenClassicStyles}
 `;
 
 StyledSplitButtonChildrenContainer.defaultProps = {